requirements.txt
discord.py==2.0.1
PlexAPI==4.0.0
fuzzywuzzy==0.18.0
pynacl==1.4.0
ffmpeg==1.4
PyYAML==5.3.1
lyricsgenius==2.0.0
